LayoutTests/inspector/debugger/regress-133182-expected.txt

CONSOLE MESSAGE: line 56: [1] Testing statement '({}).a.b.c.d;'
CONSOLE MESSAGE: line 57: [1] Paused and about to step
CONSOLE MESSAGE: line 69: [1] Resumed
CONSOLE MESSAGE: line 61: [1] Paused after stepping
CONSOLE MESSAGE: line 69: [1] Resumed
CONSOLE MESSAGE: line 1: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 56: [2] Testing statement 'exceptionBasic();'
CONSOLE MESSAGE: line 57: [2] Paused and about to step
CONSOLE MESSAGE: line 69: [2] Resumed
CONSOLE MESSAGE: line 61: [2] Paused after stepping
CONSOLE MESSAGE: line 69: [2] Resumed
CONSOLE MESSAGE: line 3: TypeError: undefined is not an object (evaluating '({}).a.b')
CONSOLE MESSAGE: line 56: [3] Testing statement 'exceptionDOM();'
CONSOLE MESSAGE: line 57: [3] Paused and about to step
CONSOLE MESSAGE: line 69: [3] Resumed
CONSOLE MESSAGE: line 61: [3] Paused after stepping
CONSOLE MESSAGE: line 69: [3] Resumed
CONSOLE MESSAGE: line 8: NotFoundError: DOM Exception 8: The object can not be found here.
CONSOLE MESSAGE: line 56: [4] Testing statement 'exceptionInHostFunction();'
CONSOLE MESSAGE: line 57: [4] Paused and about to step
CONSOLE MESSAGE: line 69: [4] Resumed
CONSOLE MESSAGE: line 61: [4] Paused after stepping
CONSOLE MESSAGE: line 69: [4] Resumed
CONSOLE MESSAGE: line 24: exception in host function
CONSOLE MESSAGE: line 56: [5] Testing statement 'throwString();'
CONSOLE MESSAGE: line 57: [5] Paused and about to step
CONSOLE MESSAGE: line 69: [5] Resumed
CONSOLE MESSAGE: line 61: [5] Paused after stepping
CONSOLE MESSAGE: line 69: [5] Resumed
CONSOLE MESSAGE: line 13: exception string
CONSOLE MESSAGE: line 56: [6] Testing statement 'throwParam({x:1});'
CONSOLE MESSAGE: line 57: [6] Paused and about to step
CONSOLE MESSAGE: line 69: [6] Resumed
CONSOLE MESSAGE: line 61: [6] Paused after stepping
CONSOLE MESSAGE: line 69: [6] Resumed
CONSOLE MESSAGE: line 18: [object Object]
CONSOLE MESSAGE: line 56: [7] Testing statement 'throwParam(new Error('error message'));'
CONSOLE MESSAGE: line 57: [7] Paused and about to step
CONSOLE MESSAGE: line 69: [7] Resumed
CONSOLE MESSAGE: line 61: [7] Paused after stepping
CONSOLE MESSAGE: line 69: [7] Resumed
CONSOLE MESSAGE: line 18: Error: error message
Regression test for https://bugs.webkit.org/show_bug.cgi?id=133182

Stepping after breaking on uncaught exceptions should not crash

PASS - paused for each uncaught exception
